% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlaps.R
\name{CalcOverlapStats}
\alias{CalcOverlapStats}
\title{Calculate Overlap Statistics}
\usage{
CalcOverlapStats(
  left,
  right,
  chr_name = "chr",
  start_name = "start",
  end_name = "end"
)
}
\arguments{
\item{left}{Data.frame containing the first set of loci.}

\item{right}{Data.frame containing the second set of loci.}

\item{chr_name}{Name of column containing chromosome.}

\item{start_name}{Name of column containing locus start position.}

\item{end_name}{Name of column containing locus end position.}
}
\value{
Data.frame.
}
\description{
Calculates overlap statistics comparing loci on the \code{left} with loci on
the \code{right}. Note this function is asymmetric;
\code{CalcOverlapStats(left, right)} will generally differ from
\code{CalcOverlapStats(right, left)}.
}
