% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/type_conversion.R
\name{DF2GR}
\alias{DF2GR}
\title{Convert from Data Frame to Genomic Ranges}
\usage{
DF2GR(data, chr_name = "chr", start_name = "start", end_name = "end")
}
\arguments{
\item{data}{Data.frame.}

\item{chr_name}{Name of column containing chromosome.}

\item{start_name}{Name of column containing locus start position.}

\item{end_name}{Name of column containing locus end position.}
}
\value{
GenomicRages object.
}
\description{
Convert from Data Frame to Genomic Ranges
}
