% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/type_conversion.R
\name{GR2DF}
\alias{GR2DF}
\title{Convert from Genomic Ranges to Data Frame}
\usage{
GR2DF(gr, chr_name = "chr", start_name = "start", end_name = "end")
}
\arguments{
\item{gr}{Genomic Ranges object.}

\item{chr_name}{Name of column containing chromosome.}

\item{start_name}{Name of column containing locus start position.}

\item{end_name}{Name of column containing locus end position.}
}
\value{
Data.frame
}
\description{
Reverses the action of \code{\link{DF2GR}}.
}
