% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\docType{class}
\name{overlap-class}
\alias{overlap-class}
\title{Overlap class}
\description{
Overlap class
}
\section{Slots}{

\describe{
\item{\code{left}}{Input loci on the left.}

\item{\code{right}}{Input loci on the right.}

\item{\code{overlaps}}{Results of \code{FindOverlaps(left, right)}.}

\item{\code{left_overlapping}}{Unique loci on the left overlapping at least 1 locus
on the right.}

\item{\code{right_overlapped}}{Unique loci on the right overlapped by at least 1
locus on the left.}

\item{\code{n_left}}{Number of input loci on the left.}

\item{\code{n_right}}{Number of input loci on the right.}
}}

