% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\name{show,overlap-method}
\alias{show,overlap-method}
\title{Show Method for Overlap Object}
\usage{
\S4method{show}{overlap}(object)
}
\arguments{
\item{object}{An object of class \code{overlap}.}
}
\description{
Show Method for Overlap Object
}
