% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlaps.R
\name{OverlapIndices}
\alias{OverlapIndices}
\title{Overlap Indices}
\usage{
OverlapIndices(
  left,
  right,
  chr_name = "chr",
  start_name = "start",
  end_name = "end"
)
}
\arguments{
\item{left}{Data.frame containing the first set of loci.}

\item{right}{Data.frame containing the second set of loci.}

\item{chr_name}{Name of column containing chromosome.}

\item{start_name}{Name of column containing locus start position.}

\item{end_name}{Name of column containing locus end position.}
}
\value{
Data.frame.
}
\description{
Identifies which loci from \code{left} overlap with which
loci from \code{right}.
}
