% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\name{print.overlap}
\alias{print.overlap}
\title{Print Method for Overlap Object.}
\usage{
\method{print}{overlap}(x, ...)
}
\arguments{
\item{x}{An object of class \code{overlap}.}

\item{...}{Unused.}
}
\description{
Print method for objects of class \code{overlap}.
}
