% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlaps.R
\name{OverlapIndices}
\alias{OverlapIndices}
\title{Overlap Indices}
\usage{
OverlapIndices(
  left,
  right,
  left_chr_name = "chr",
  left_start_name = "start",
  left_end_name = "end",
  right_chr_name = "chr",
  right_start_name = "start",
  right_end_name = "end"
)
}
\arguments{
\item{left}{Data.frame containing the first set of loci.}

\item{right}{Data.frame containing the second set of loci.}

\item{left_chr_name}{Name of column containing chromosome.}

\item{left_start_name}{Name of column containing locus start position.}

\item{left_end_name}{Name of column containing locus end position.}

\item{right_chr_name}{Name of column containing chromosome.}

\item{right_start_name}{Name of column containing locus start position.}

\item{right_end_name}{Name of column containing locus end position.}
}
\value{
Data.frame.
}
\description{
Identifies which loci from \code{left} overlap with which
loci from \code{right}.
}
